+++ app/eilmer3/source/bc_fstc.hh
@@ -6,7 +6,7 @@ class fstcBC : public BoundaryCondition {
 private:
     std::string filename;
     std::vector<double> fstc_TProfile;
-    int ncell_for_profile;
+    size_t ncell_for_profile;
 public:
     fstcBC( Block *bdp, int which_boundary, const std::string filename="fstc_temp.txt" );
     fstcBC( const fstcBC &bc );
