+++ app/eilmer3/source/bc_static_profile.hh
@@ -5,13 +5,13 @@
 class StaticProfileBC : public BoundaryCondition {
 private:
     std::string filename;
-    int n_profile;
-    int nsp, nmodes;
-    int ncell_for_profile;
+    size_t n_profile;
+    size_t nsp, nmodes;
+    size_t ncell_for_profile;
     std::vector<CFlowCondition*> flow_profile;
 public:
     StaticProfileBC( Block *bdp, int which_boundary, 
-		     const std::string filename="profile.dat", int n_profile=1 );
+		     const std::string filename="profile.dat", size_t n_profile=1 );
     StaticProfileBC( const StaticProfileBC &bc );
     StaticProfileBC();
     StaticProfileBC & operator=(const StaticProfileBC &bc);
